Renamed the Connection object to be VestPocketStore, to better separate it from other connection type objects (as it is not an ADO.NET data provider) and added an VestPocketOptions parameter to hold configuration choices for the store instance. This should make registering a VestPocketStore instance easier (such as when binding appsettings.json config values to an options instance).
Initiation was also moved to the constructor, and completely separated from the method of opening the store (instead of having an internal constructor and static methods to create an instance). A validation method was added for the VestPocketOptions class.

A ReadOnly mode was also added, which when used can ellide the read locks, almost doubling read performance.

The README.md file was also updated to reflect these changes.