Flushing background task: Task was replaced with an explicit thread to avoid allocations from Task.Delay.
GetNextLineFeedPosition was made synchronous to avoid state machine generation while awaiting the result. This method is used from a single background Task from each VestPocketStore, and in tight loops during loading a store.

Also added benchmarks to show performance of instantiating an entity, as well as setting 10 entities in a single transaction.